+++ package.json
@@ -69,7 +69,6 @@
     "build:stark-ui": "npm run build -- --packages=stark-ui",
     "build:showcase": "cd showcase && npm run build:prod && cd ..",
     "build:showcase:ghpages": "cd showcase && npm run build:prod:ghpages && cd ..",
-    "build:showcase:ghpages:ci-test-env": "cd showcase && npm run build:prod:ghpages:ci-test-env && cd ..",
     "build:starter": "cd starter && npm run build && cd ..",
     "check:starter:stark-versions": "node ./check-stark-versions.js ./starter/package.json latest",
     "clean": "npx rimraf ./dist",
@@ -145,7 +144,6 @@
     "test:showcase:e2e:browserstack": "cd showcase && npm run e2e:browserstack && cd ../..",
     "test:starter": "cd starter && npm run test-fast && cd ../..",
     "test:starter:e2e": "cd starter && npm run e2e && cd ../..",
-    "test:starter:e2e:browserstack": "cd starter && npm run e2e:browserstack && cd ../..",
     "test:ci:all": "npm run test:ci:stark-core && npm run test:ci:stark-ui && npm run test:ci:starter && npm run test:ci:showcase",
     "test:ci:stark-core": "cd packages/stark-core && npm run test-fast:ci && cd ../..",
     "test:ci:stark-ui": "cd packages/stark-ui && npm run test-fast:ci && cd ../..",
