+++ packages/stark-build/package.json
@@ -19,55 +19,29 @@
   },
   "types": "typings/index.d.ts",
   "dependencies": {
+    "@angular-builders/custom-webpack": "^7.3.1",
+    "@angular-builders/dev-server": "^7.3.1",
     "@angular-devkit/build-angular": "^0.13.0",
     "@babel/core": "^7.2.2",
     "@babel/preset-env": "^7.2.3",
-    "@types/uglify-js": "^3.0.4",
     "@types/webpack": "^4.4.24",
-    "add-asset-html-webpack-plugin": "^3.1.2",
-    "angular-named-lazy-chunks-webpack-plugin": "^2.1.0",
-    "assets-webpack-plugin": "^3.9.7",
     "babel-loader": "^8.0.5",
     "base-href-webpack-plugin": "^2.0.0",
-    "circular-dependency-plugin": "^5.0.2",
     "codelyzer": "^4.5.0",
-    "copy-webpack-plugin": "^4.6.0",
-    "css-loader": "^2.1.0",
-    "file-loader": "^3.0.1",
+    "ejs-loader": "^0.3.1",
     "html-elements-webpack-plugin": "^2.0.0",
     "html-webpack-plugin": "^3.2.0",
-    "mini-css-extract-plugin": "^0.5.0",
-    "node-sass": "^4.11.0",
-    "optimize-css-assets-webpack-plugin": "^5.0.1",
-    "postcss-cssnext": "^3.1.0",
-    "postcss-import": "^12.0.1",
-    "postcss-loader": "^3.0.0",
-    "postcss-nesting": "^7.0.0",
-    "postcss-simple-extend": "^1.0.0",
-    "postcss-url": "^8.0.0",
-    "preload-webpack-plugin": "^2.3.0",
     "prettier": "^1.16.1",
-    "raw-loader": "^1.0.0",
-    "sass-loader": "^7.1.0",
-    "source-map-loader": "^0.2.4",
-    "string-replace-loader": "^2.1.1",
-    "style-loader": "^0.23.1",
     "stylelint": "^9.10.1",
     "stylelint-config-prettier": "^4.0.0",
     "stylelint-webpack-plugin": "^0.10.5",
-    "to-string-loader": "^1.1.5",
     "tslint": "^5.11.0",
     "tslint-config-prettier": "^1.17.0",
     "tslint-loader": "^3.6.0",
     "tslint-sonarts": "^1.8.0",
-    "uglifyjs-webpack-plugin": "^2.1.0",
-    "url-loader": "^1.1.2",
-    "webpack": "^4.29.0",
-    "webpack-cli": "^3.2.0",
-    "webpack-dev-server": "^3.1.11",
+    "webpack-bundle-analyzer": "^3.0.4",
     "webpack-merge": "^4.2.1",
-    "webpack-monitor": "^1.0.14",
-    "write-file-webpack-plugin": "^4.5.0"
+    "webpack-monitor": "^1.0.14"
   },
   "devDependencies": {
     "@angular/cli": "^7.0.0",
