+++ packages/stark-ui/src/modules/table/components/_table.component.scss
@@ -55,7 +55,7 @@
           display: flex;
           flex-direction: row;
           align-items: center;
-          justify-content: start;
+          justify-content: flex-start;
 
           .sort-header {
             align-items: center;
