+++ showcase/package.json
@@ -26,16 +26,16 @@
     "npm": ">=5.6.0"
   },
   "scripts": {
-    "build:aot:prod": "npm run clean:dist && npm run clean:aot && cross-env BUILD_AOT=1 npm run webpack -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.prod.js  --progress --profile --bail",
-    "build:aot:prod:ci-test-env": "npm run clean:dist && npm run clean:aot && cross-env BUILD_AOT=1 npm run webpack -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.prod.js  --progress --profile --bail --ci-test-env",
+    "build:aot:prod": "npm run clean:dist && npm run clean:aot && npm run ng -- build --prod",
     "build:aot": "npm run build:aot:prod",
-    "build:aot:dev": "cross-env BUILD_AOT=1 npm run build:dev",
-    "build:dev": "npm run clean:dist && npm run webpack -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.dev.js --progress --profile",
-    "build:dev:monitor": "npx mkdirp reports && npm run build:dev -- --env.monitor",
+    "build:aot:dev": "npm run build:dev -- --aot",
+    "build:dev": "npm run clean:dist && npm run ng -- build",
+    "build:dev:monitor": "npx mkdirp reports && cross-env MONITOR=1 npm run build:dev",
+    "build:dev:bundle-analyzer": "npx mkdirp reports && cross-env BUNDLE_ANALYZER=1 npm run build:dev",
     "build:docker": "npm run build:prod && docker build -t angular2-webpack-start:latest .",
     "build:prod": "npm run clean:dist && npm run build:aot:prod",
+    "build:prod:bundle-analyzer": "npm run clean:dist && cross-env BUNDLE_ANALYZER=1 npm run build:aot:prod",
     "build:prod:ghpages": "npm run clean:dist && node ./ghpages-adapt-angular-json.js && npm run build:aot:prod",
-    "build:prod:ghpages:ci-test-env": "npm run clean:dist && node ./ghpages-adapt-angular-json.js && npm run build:aot:prod:ci-test-env",
     "build": "npm run build:dev",
     "check-deps": "npx npm-check-u",
     "ci:aot": "cross-env BUILD_E2E=1 npm run lint && npm run test && npm run build:aot && npm run e2e",
@@ -62,6 +62,7 @@
     "lint": "npm run lint-ts && npm run lint-css",
     "lint-ts": "tslint --config ./tslint.json --project ./tsconfig.json --format codeFrame",
     "lint-css": "stylelint \"./src/**/*.?(pc|sc|c|sa)ss\" --formatter \"string\"",
+    "ng": "node --max_old_space_size=4096 ./node_modules/@angular/cli/bin/ng",
     "ngc": "ngc",
     "node": "node",
     "postinstall": "npm run webdriver:update",
@@ -75,14 +76,16 @@
     "protractor:delay": "sleep 3 && npm run protractor",
     "protractor:live": "protractor --elementExplorer",
     "rimraf": "npx rimraf",
-    "server:dev:hmr": "npm run server:dev -- --hot",
-    "server:aot:dev": "cross-env BUILD_AOT=1 npm run server:dev",
-    "server:dev": "npm run clean:dist && npm run webpack-dev-server -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.dev.js --open --progress --profile --watch --content-base src/",
+    "server:dev:hmr": "npm run server:dev -- --configuration hmr",
+    "server:aot:dev": "npm run server:dev -- --aot",
+    "server:dev": "npm run clean:dist && npm run ng -- serve",
+    "server:dev:monitor": "npm run clean:dist && cross-env MONITOR=1 npm run ng -- serve",
+    "server:dev:bundle-analyzer": "npm run clean:dist && cross-env BUNDLE_ANALYZER=1 npm run ng -- serve",
     "server:prod": "http-server dist -c-1 --cors",
     "server:prod:ci": "http-server dist -p 3000 -c-1 --cors",
     "server": "npm run server:dev",
     "start:hmr": "npm run server:dev:hmr",
-    "start:monitor": "npx mkdirp reports && npm run server:dev -- --env.monitor",
+    "start:monitor": "npx mkdirp reports && cross-env MONITOR=1 npm run server:dev",
     "start": "npm run server:dev",
     "start:aot": "npm run server:aot:dev",
     "test": "npm run lint && npm run test-fast",
@@ -120,8 +123,8 @@
     "@angular/platform-browser-dynamic": "~7.2.2",
     "@angular/platform-server": "~7.2.2",
     "@angular/router": "~7.2.2",
-    "@nationalbankbelgium/stark-core": "file:../dist/packages-dist/stark-core/nationalbankbelgium-stark-core-10.0.0-beta.2-ebdff291.tgz",
-    "@nationalbankbelgium/stark-ui": "file:../dist/packages-dist/stark-ui/nationalbankbelgium-stark-ui-10.0.0-beta.2-ebdff291.tgz",
+    "@nationalbankbelgium/stark-core": "file:../dist/packages-dist/stark-core/nationalbankbelgium-stark-core-10.0.0-beta.4-a03d1727.tgz",
+    "@nationalbankbelgium/stark-ui": "file:../dist/packages-dist/stark-ui/nationalbankbelgium-stark-ui-10.0.0-beta.4-a03d1727.tgz",
     "@uirouter/visualizer": "~7.0.0",
     "angular-in-memory-web-api": "~0.8.0",
     "basscss": "~8.0.10",
@@ -137,8 +140,8 @@
   },
   "devDependencies": {
     "@compodoc/compodoc": "~1.1.7",
-    "@nationalbankbelgium/stark-build": "file:../dist/packages-dist/stark-build/nationalbankbelgium-stark-build-10.0.0-beta.2-ebdff291.tgz",
-    "@nationalbankbelgium/stark-testing": "file:../dist/packages-dist/stark-testing/nationalbankbelgium-stark-testing-10.0.0-beta.2-ebdff291.tgz",
+    "@nationalbankbelgium/stark-build": "file:../dist/packages-dist/stark-build/nationalbankbelgium-stark-build-10.0.0-beta.4-a03d1727.tgz",
+    "@nationalbankbelgium/stark-testing": "file:../dist/packages-dist/stark-testing/nationalbankbelgium-stark-testing-10.0.0-beta.4-a03d1727.tgz",
     "@types/core-js": "~2.5.0",
     "@types/hammerjs": "~2.0.36",
     "@types/node": "~8.10.37",
