+++ starter/angular.json
@@ -9,8 +9,17 @@
       "projectType": "application",
       "architect": {
         "build": {
-          "builder": "@angular-devkit/build-angular:browser",
+          "builder": "@angular-builders/custom-webpack:browser",
           "options": {
+            "customWebpackConfig": {
+              "path": "./node_modules/@nationalbankbelgium/stark-build/config/webpack-partial.dev.js",
+              "mergeStrategies": {
+                "modules.rules": "prepend",
+                "plugins": "prepend",
+                "devServer": "prepend",
+                "replaceDuplicatePlugins": false
+              }
+            },
             "outputPath": "dist",
             "index": "src/index.html",
             "main": "src/main.browser.ts",
@@ -19,23 +28,33 @@
             "assets": [
               {
                 "glob": "mdi.svg",
-                "input": "../node_modules/@mdi/angular-material",
-                "output": "./assets/material-icons"
+                "input": "node_modules/@mdi/angular-material",
+                "output": "assets/material-icons"
               },
               {
                 "glob": "**/*",
-                "input": "assets",
-                "output": "./assets"
+                "input": "src/assets",
+                "output": "assets"
               },
               {
                 "glob": "**/*",
-                "input": "assets-base",
-                "output": "./"
+                "input": "src/assets-base",
+                "output": "."
+              },
+              {
+                "glob": "**/*",
+                "input": "node_modules/@nationalbankbelgium/stark-core/assets",
+                "output": "assets/stark-core"
+              },
+              {
+                "glob": "**/*",
+                "input": "node_modules/@nationalbankbelgium/stark-ui/assets",
+                "output": "assets/stark-ui"
               }
             ],
             "styles": ["node_modules/@nationalbankbelgium/stark-core/assets/css/pre-load-style.css", "src/styles/styles.scss"],
             "scripts": [],
-            "deployUrl": "",
+            "deployUrl": "/",
             "baseHref": "/"
           },
           "configurations": {
@@ -48,11 +67,18 @@
               ]
             },
             "production": {
+              "customWebpackConfig": {
+                "path": "./node_modules/@nationalbankbelgium/stark-build/config/webpack-partial.prod.js",
+                "mergeStrategies": {
+                  "modules.rules": "prepend",
+                  "plugins": "prepend",
+                  "replaceDuplicatePlugins": false
+                }
+              },
               "optimization": true,
               "outputHashing": "all",
               "sourceMap": false,
               "extractCss": true,
-              "namedChunks": false,
               "aot": true,
               "extractLicenses": true,
               "vendorChunk": false,
@@ -69,7 +95,6 @@
               "outputHashing": "all",
               "sourceMap": false,
               "extractCss": true,
-              "namedChunks": false,
               "aot": true,
               "extractLicenses": true,
               "vendorChunk": false,
@@ -84,12 +109,15 @@
           }
         },
         "serve": {
-          "builder": "@angular-devkit/build-angular:dev-server",
+          "builder": "@angular-builders/dev-server:generic",
           "options": {
-            "browserTarget": "starter:build"
+            "browserTarget": "starter:build",
+            "port": 3000,
+            "open": true
           },
           "configurations": {
             "hmr": {
+              "hmr": true,
               "browserTarget": "starter:build:hmr"
             },
             "production": {
@@ -114,17 +142,41 @@
           }
         },
         "test": {
+          "builder": "@angular-devkit/build-angular:karma",
           "options": {
             "main": "base.spec.ts",
+            "karmaConfig": "./karma.conf.js",
             "tsConfig": "tsconfig.spec.json"
           }
         }
       }
     },
     "starter-e2e": {
-      "root": "",
+      "root": "e2e/",
       "sourceRoot": "",
-      "projectType": "application"
+      "projectType": "application",
+      "prefix": "",
+      "architect": {
+        "e2e": {
+          "builder": "@angular-devkit/build-angular:protractor",
+          "options": {
+            "protractorConfig": "e2e/protractor.conf.js",
+            "devServerTarget": "starter:serve"
+          },
+          "configurations": {
+            "production": {
+              "devServerTarget": "starter:serve:production"
+            }
+          }
+        },
+        "lint": {
+          "builder": "@angular-devkit/build-angular:tslint",
+          "options": {
+            "tsConfig": "e2e/tsconfig.e2e.json",
+            "exclude": ["**/node_modules/**"]
+          }
+        }
+      }
     }
   },
   "defaultProject": "starter",
