+++ starter/package.json
@@ -26,13 +26,15 @@
     "npm": ">=5.6.0"
   },
   "scripts": {
-    "build:aot:prod": "npm run clean:dist && npm run clean:aot && cross-env BUILD_AOT=1 npm run webpack -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.prod.js  --progress --profile --bail",
+    "build:aot:prod": "npm run clean:dist && npm run clean:aot && npm run ng -- build --prod",
     "build:aot": "npm run build:aot:prod",
-    "build:aot:dev": "cross-env BUILD_AOT=1 npm run build:dev",
-    "build:dev": "npm run clean:dist && npm run webpack -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.dev.js --progress --profile",
-    "build:dev:monitor": "npx mkdirp reports && npm run build:dev -- --env.monitor",
+    "build:aot:dev": "npm run build:dev -- --aot",
+    "build:dev": "npm run clean:dist && npm run ng -- build",
+    "build:dev:monitor": "npx mkdirp reports && cross-env MONITOR=1 npm run build:dev",
+    "build:dev:bundle-analyzer": "npx mkdirp reports && cross-env BUNDLE_ANALYZER=1 npm run build:dev",
     "build:docker": "npm run build:prod && docker build -t angular2-webpack-start:latest .",
     "build:prod": "npm run clean:dist && npm run build:aot:prod",
+    "build:prod:bundle-analyzer": "npm run clean:dist && cross-env BUNDLE_ANALYZER=1 npm run build:aot:prod",
     "build": "npm run build:dev",
     "check-deps": "npx npm-check-u",
     "ci:aot": "cross-env BUILD_E2E=1 npm run lint && npm run test && npm run build:aot && npm run e2e",
@@ -61,6 +63,7 @@
     "lint": "npm run lint-ts && npm run lint-css",
     "lint-ts": "tslint --config ./tslint.json --project ./tsconfig.json --format codeFrame",
     "lint-css": "stylelint \"./src/**/*.?(pc|sc|c|sa)ss\" --formatter \"string\"",
+    "ng": "node --max_old_space_size=4096 ./node_modules/@angular/cli/bin/ng",
     "ngc": "ngc",
     "node": "node",
     "postinstall": "npm run webdriver:update",
@@ -73,14 +76,16 @@
     "protractor:delay": "sleep 3 && npm run protractor",
     "protractor:live": "protractor --elementExplorer",
     "rimraf": "npx rimraf",
-    "server:dev:hmr": "npm run server:dev -- --hot",
-    "server:aot:dev": "cross-env BUILD_AOT=1 npm run server:dev",
-    "server:dev": "npm run clean:dist && npm run webpack-dev-server -- --config node_modules/@nationalbankbelgium/stark-build/config/webpack.dev.js --open --progress --profile --watch --content-base src/",
+    "server:dev:hmr": "npm run server:dev -- --configuration hmr",
+    "server:aot:dev": "npm run server:dev -- --aot",
+    "server:dev": "npm run clean:dist && npm run ng -- serve",
+    "server:dev:monitor": "npm run clean:dist && cross-env MONITOR=1 npm run ng -- serve",
+    "server:dev:bundle-analyzer": "npm run clean:dist && cross-env BUNDLE_ANALYZER=1 npm run ng -- serve",
     "server:prod": "http-server dist -c-1 --cors",
     "server:prod:ci": "http-server dist -p 3000 -c-1 --cors",
     "server": "npm run server:dev",
     "start:hmr": "npm run server:dev:hmr",
-    "start:monitor": "npx mkdirp reports && npm run server:dev -- --env.monitor",
+    "start:monitor": "npx mkdirp reports && cross-env MONITOR=1 npm run server:dev",
     "start": "npm run server:dev",
     "start:aot": "npm run server:aot:dev",
     "test": "npm run lint && npm run test-fast",
